add zoom feature

+++ collaboration-suite-core-ui/src/main/java/com/kaleyra/collaboration_suite_core_ui/call/CallUIController.kt
@@ -93,7 +93,7 @@ interface CallUIController {
      *
      * @param value zoom level
      */
-    fun onSetZoom(value: Int)
+    fun onSetZoom(value: Float)
 }
 
 //interface CallWaitingUIController {
